+++ modules/Bot/Bot.js
@@ -2,8 +2,6 @@ const puppeteer = require('puppeteer');
 
 class Bot{
 
- 
-
     async newBrowser(param){
         const browser = await puppeteer.launch( {headless:param, ignoreHTTPSErrors: true} );
         this.browser = browser
@@ -18,17 +16,9 @@ class Bot{
     async goTo(url){
         await this.page.goto(url,{timeout:10000});
     }
-
-    async scraping(){
-    const log = await this.page.evaluate(async() =>{
-            return {
-                
-                title: document.querySelector('title').innerHTML, 
-                
-                }
-            }
-        );
-        return log
+    
+    async closeBrowser(){
+        this.browser.close()
     }
 }
